Reduce a bit of code duplication
