+++ crates/subspace-node/src/lib.rs
@@ -20,8 +20,28 @@ mod chain_spec;
 
 use clap::Parser;
 use sc_cli::SubstrateCli;
-use sc_executor::RuntimeVersion;
+use sc_executor::{NativeExecutionDispatch, RuntimeVersion};
 use sc_service::ChainSpec;
+
+/// Executor dispatch for subspace runtime
+pub struct ExecutorDispatch;
+
+impl NativeExecutionDispatch for ExecutorDispatch {
+    /// Only enable the benchmarking host functions when we actually want to benchmark.
+    #[cfg(feature = "runtime-benchmarks")]
+    type ExtendHostFunctions = frame_benchmarking::benchmarking::HostFunctions;
+    /// Otherwise we only use the default Substrate host functions.
+    #[cfg(not(feature = "runtime-benchmarks"))]
+    type ExtendHostFunctions = ();
+
+    fn dispatch(method: &str, data: &[u8]) -> Option<Vec<u8>> {
+        subspace_runtime::api::dispatch(method, data)
+    }
+
+    fn native_version() -> sc_executor::NativeVersion {
+        subspace_runtime::native_version()
+    }
+}
 
 /// Utilities for working with a node.
 #[derive(Debug, clap::Subcommand)]
